requirements.txt
# Liquidation Blitz - Python Dependencies

# Streamlit UI
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # Excel file support

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0  # Image handling for PDFs

# AWS S3
boto3>=1.28.0
botocore>=1.31.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests (for downloading images)
requests>=2.31.0
